Deleted concluded

+++ contacts/src/app/contact/listagem/listagem.component.ts
@@ -1,5 +1,7 @@
-import { Component, OnInit, Input, Output } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
+import { ErrorConfirmationComponent } from './../../shared/components/error-confirmation/error-confirmation.component';
+import { Component, OnInit, Input, Output, Inject } from '@angular/core';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { catchError, Observable, of } from 'rxjs';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 import { ModalComponent } from 'src/app/shared/components/modal/modal.component';
@@ -23,9 +25,13 @@ export class ListagemComponent implements OnInit {
   displayedColumns = ['id', 'name', 'lastName', 'email', 'phone', 'actions'];
   //contacts: Contact[] = []; // lista de atributo de pessoas
 
-  constructor(private SoluctionService: SolutionService, //adicionando no construtor a injeção de serviço
+  constructor(
+    private SoluctionService: SolutionService, //adicionando no construtor a injeção de serviço
     public dialog: MatDialog,
-    public modalDialog: MatDialog
+    public modalDialog: MatDialog,
+    private snack: MatSnackBar
+
+    // public dialogRef: MatDialogRef<ModalComponent>
   ) { }
 
   ngOnInit(): void {
@@ -41,14 +47,6 @@ export class ListagemComponent implements OnInit {
         })
       );
   }
-
-  // save(data: Contact) {
-  //   this.SoluctionService.save(data)
-  //     .subscribe(() => {
-  //       this.update()
-  //     });
-  //     console.log('atualizou');
-  // }
 
   onError(errorMsg: string) {
     this.dialog.open(ErrorDialogComponent, {
@@ -66,16 +64,26 @@ export class ListagemComponent implements OnInit {
       });
   }
 
-  removeContact(contact: Contact) {
-    this.SoluctionService.delete(contact).subscribe(() => {
-      this.updateContact();
+  delete(contact: Contact) {
+    const dialogRef = this.dialog.open(ErrorConfirmationComponent, {
+      width: '250px',
+      data: { nome: contact.name } // Envia o nome do contato para o diálogo de confirmação
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.SoluctionService.delete(contact).subscribe(() => {
+          this.updateContact();
+          this.onSuccessRemove();
+        });
+      }
     });
   }
-}
 
-  // onEdit(contact: Contact) {
-  //   this.edit.emit(contact);
-  // }
+  onSuccessRemove() {
+    this.snack.open('Contato Removido com sucesso!', '', { duration: 5000 });
+  }
+}
